+++ examples/ui/ethereum.world/public/index.html
@@ -25,28 +25,53 @@
     <link rel="stylesheet" href="/fonts/inter.css" />
     <title>Ethereum World</title>
     <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/systemjs/6.7.1/system.min.js"
+      integrity="sha512-R1dzyrUOrbjCRxsRK7JRUVJA5yFEHQ9Ui/qcyon6QRbcVeSzz1SfHzqVuVZh9CPGiCftiJpeHOLz+uFcVNevYQ=="
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/systemjs/6.7.1/extras/amd.min.js"
+      integrity="sha512-fRfysEnVm90UgqFUsB+UlgnHK5vvSnsl64hxHaJJl4wE+P1sTwDUuQGmohtg/CSK60mY5A4elesW6UWQdLQsAg=="
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/systemjs/6.7.1/extras/named-exports.min.js"
+      integrity="sha512-mf4/y3IroNfkl20Brygdc+tb4a38MrEmU+VTSyMaqlsZhfVHCdHniA//HEgETGlJptSSCPv72Dy6kXB+OHiGAQ=="
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/systemjs/6.7.1/extras/use-default.min.js"
+      integrity="sha512-Ahpc6VqzcdPS1xTuBGdnw+l64Iz6KNXAy/4/1bBzAv6nTl+NK7mnYHMDtI8wAzd2PTV9GVNURwKf87y36eWZ+g=="
+      crossorigin="anonymous"
+    ></script>
+    <meta name="importmap-type" content="systemjs-importmap" />
+    <script type="systemjs-importmap" src="/importmap.json"></script>
+
+    <link rel="preload" href="https://unpkg.com/react@17/umd/react.development.js" as="script" crossorigin></link>
+    <link rel="preload" href="https://unpkg.com/react-dom@17/umd/react-dom.development.js" as="script" crossorigin></link>
+    <!-- <script
       src="https://cdnjs.cloudflare.com/ajax/libs/react/17.0.2/umd/react.production.min.js"
       integrity="sha512-qlzIeUtTg7eBpmEaS12NZgxz52YYZVF5myj89mjJEesBd/oE9UPsYOX2QAXzvOAZYEvQohKdcY8zKE02ifXDmA=="
       crossorigin="anonymous"
-    ></script>
-    <script
+    ></script> -->
+    <!-- <script
       src="https://cdnjs.cloudflare.com/ajax/libs/react-is/17.0.2/umd/react-is.production.min.js"
       integrity="sha512-7jhr1ZRXGaW9PsKijm984f52EsZ16CalG0EnpM7VUZmxMULSRMqaRs7l6Rumx/0gzgQiB7ZBkS5pdzDwDR56cQ=="
       crossorigin="anonymous"
-    ></script>
-    <script
+    ></script> -->
+    <!-- <script
       src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/17.0.2/umd/react-dom.production.min.js"
       integrity="sha512-9jGNr5Piwe8nzLLYTk8QrEMPfjGU0px80GYzKZUxi7lmCfrBjtyCc1V5kkS5vxVwwIB7Qpzc7UxLiQxfAN30dw=="
       crossorigin="anonymous"
-    ></script>
+    ></script> -->
     <!-- <script src="https://unpkg.com/single-spa-react@3.2.0/lib/umd/single-spa-react.js"></script> -->
-    <script
+    <!-- <script
       src="https://cdnjs.cloudflare.com/ajax/libs/styled-components/5.2.3/styled-components.min.js"
       integrity="sha512-B9cYmQHqdZ45bpfpSO5IC3UN9hIfQe7/bmiG2aXXeakS8uB3hVK1NomxvcbZoEkZpRt6iGASR++dqfd/+f7OSQ=="
       crossorigin="anonymous"
-    ></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/rxjs/6.6.7/rxjs.umd.min.js" integrity="sha512-0/2ebe9lI6BcinFBXFjbBkquDfccT2wP+E48pecciFuGMXPRNdInXZawHiM2NUUVJ4/aKAzyebbvh+CkvRhwTA==" crossorigin="anonymous"></script>
-    <script src="/design/main.js" async></script>
+    ></script> -->
+    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/rxjs/6.6.7/rxjs.umd.min.js" integrity="sha512-0/2ebe9lI6BcinFBXFjbBkquDfccT2wP+E48pecciFuGMXPRNdInXZawHiM2NUUVJ4/aKAzyebbvh+CkvRhwTA==" crossorigin="anonymous"></script> -->
+    <!-- <script src="/design/main.js" async></script> -->
     <style>
       html,
       body,
