+++ examples/ui/ethereum.world/webpack.config.js
@@ -1,6 +1,4 @@
-const webpack = require('webpack');
 const CopyPlugin = require('copy-webpack-plugin');
-const { ModuleFederationPlugin } = webpack.container;
 const path = require('path');
 const baseConfig = require('../../../ui/webpack.config');
 
@@ -11,20 +9,6 @@ module.exports = Object.assign(baseConfig, {
     path: path.resolve(__dirname, 'dist'),
   },
   plugins: baseConfig.plugins.concat([
-    new ModuleFederationPlugin({
-      name: 'ethereum.world',
-      shared: {
-        react: {
-          singleton: true,
-        },
-        'react-dom': {
-          singleton: true,
-        },
-        'styled-components': {
-          singleton: true,
-        },
-      },
-    }),
     new CopyPlugin({
       patterns: [
         { from: path.resolve(__dirname, '../../../ui/build') },
