+++ examples/ui/feed-app/src/index.ts
@@ -107,5 +107,6 @@
   //     area: MenuItemAreaType.AppArea,
   //   },
   // });
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
 })(window.System);
