+++ ui/design/package.json
@@ -68,7 +68,8 @@
     "react-fast-compare": "3.2.0",
     "react-virtualized-auto-sizer": "1.0.2",
     "react-toast-notifications": "2.4.0",
-    "react-transition-group": "4.4.1"
+    "react-transition-group": "4.4.1",
+    "styled-components": "5.2.3"
   },
   "devDependencies": {
     "@akashaproject/ui-awf-typings": "*",
