+++ ui/design/webpack.config.js
@@ -1,11 +1,15 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require('path');
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const baseConfig = require('../webpack.config');
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const packageName = require('./package.json').name;
 
 module.exports = Object.assign(baseConfig, {
   context: path.resolve(__dirname),
-  output: {
+  output: Object.assign(baseConfig.output, {
     library: packageName.replace(/@/, '').replace(/\//, '__').replace(/-/, '_'),
-    libraryTarget: 'umd',
-  },
+    filename: 'index.js',
+    publicPath: '/design/'
+  }),
 });
