--- ui/plugins/profile/src/index.ts
@@ -1,4 +1,0 @@
-// @ts-ignore
-import bootstrap from './bootstrap';
-
-bootstrap(() => {});
