+++ ui/webpack.config.js
@@ -1,5 +1,8 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const webpack = require('webpack');
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const commons = require('./app.pack.conf');
 
 const isProduction = process.env.NODE_ENV === 'production';
@@ -23,6 +26,7 @@ module.exports = {
     extensions: ['.ts', '.tsx', '.js'],
   },
   output: {
+    libraryTarget: 'umd',
     publicPath: 'auto',
   },
   plugins: [
@@ -31,7 +35,7 @@ module.exports = {
       MODERATION_API: process.env.MODERATION_API || 'https://staging-moderation.ethereum.world',
       INFURA_ID: process.env.INFURA_ID || '',
       BUCKET_VERSION: process.env.BUCKET_VERSION || '',
-      EWA_MAILSENDER: process.env.EWA_MAILSENDER || ''
+      EWA_MAILSENDER: process.env.EWA_MAILSENDER || '',
     }),
     new webpack.AutomaticPrefetchPlugin(),
     new webpack.ProgressPlugin({
