--- ui/widgets/login/src/index.ts
@@ -1,5 +1,0 @@
-// @ts-ignore
-import bootstrap from './bootstrap';
-bootstrap(() => {
-  //
-});
