+++ ui/widgets/login/webpack.config.js
@@ -1,29 +1,16 @@
 const path = require('path');
-const ModuleFederationPlugin = require('webpack/lib/container/ModuleFederationPlugin');
 const baseConfig = require('../../webpack.config');
 const packageName = require('./package.json').name;
 
-module.exports = Object.assign(baseConfig, {
-  context: path.resolve(__dirname),
-  plugins: baseConfig.plugins.concat([
-    new ModuleFederationPlugin({
-      // akashaproject__ui_widget_login
-      name: packageName.replace(/@/g, '').replace(/\//g, '__').replace(/-/g, '_'),
+const config = {
+    context: path.resolve(__dirname),
+    output: {
+      libraryTarget: baseConfig.output.libraryTarget,
+      library: packageName.replace(/@/g, '').replace(/\//g, '__').replace(/-/g, '_'),
+      path: path.resolve(__dirname, 'dist'),
       filename: 'index.js',
-      exposes: {
-        './app': './src/bootstrap',
-      },
-      shared: {
-        react: {
-          singleton: true,
-        },
-        'react-dom': {
-          singleton: true,
-        },
-        'styled-components': {
-          singleton: true,
-        },
-      },
-    }),
-  ]),
-});
+      publicPath: '/widgets/',
+    },
+  };
+
+module.exports = Object.assign({}, baseConfig, config);
