--- ui/widgets/sidebar/src/index.ts
@@ -1,6 +1,0 @@
-// @ts-ignore
-import bootstrap from './bootstrap';
-
-bootstrap(() => {
-  //
-});
