+++ ui/widgets/trending/src/index.tsx
